requirements.txt
# NeuralSite Core Dependencies

# Web Framework
fastapi>=0.100
uvicorn>=0.23

# Data Validation
pydantic>=2.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0
neo4j>=5.0

# Data Processing
numpy>=1.24
pandas>=2.0
openpyxl>=3.1

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9
geoalchemy2>=0.14

# Environment
python-dotenv>=1.0

# Optional: BIM Export
# ifcopenshell>=0.7  # For IFC format export
